+++ blog_finder/DatabaseManager.py
