+++ blog_finder/settings.py
